Move secrets mount path to /var/openfaas/secrets
**What**
- change the secret configuration string spec to mount secrets in
/var/openfaas/secrets

**Why**
- this will ensure that the secrets required for the function business
logic do not interfer with standard locations that other systems may
want to modify.
- this ensures a consistent experience for function deployment across
swarm and k8s

Implements openfaas/faas#692
Relates to openfaas/faas-netes#188

**Notes**
